+++ public_transport_watcher/db/migrations/alembic/versions/e0deb01a69ec_addind_a_cat_day_column_to_the_.py
@@ -1,4 +1,4 @@
-"""addind a cat_day column to the transport time_bin table
+"""adding a cat_day column to the transport time_bin table
 
 Revision ID: e0deb01a69ec
 Revises: 58cdd1caec62
@@ -11,6 +11,7 @@ from typing import Sequence, Union
 from alembic import op
 import sqlalchemy as sa
 
+from public_transport_watcher.db.models.enums import DayCategoryEnum
 
 # revision identifiers, used by Alembic.
 revision: str = "e0deb01a69ec"
@@ -21,9 +22,23 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade() -> None:
     """Upgrade schema."""
+    enum_values = [e.value for e in DayCategoryEnum]
+    enum_name = "cat_day"
+    enum_schema = "transport"
+
+    op.execute(
+        f"CREATE TYPE {enum_schema}.{enum_name} AS ENUM ({', '.join(repr(v) for v in enum_values)})"
+    )
+
     op.add_column(
         "time_bin", sa.Column("cat_day", sa.String(), nullable=True), schema="transport"
     )
+    op.execute("""
+        ALTER TABLE transport.time_bin 
+        ALTER COLUMN cat_day TYPE transport.cat_day 
+        USING cat_day::text::transport.cat_day
+    """)
+
     op.drop_column("station", "wording", schema="weather")
     op.drop_column("station", "wording", schema="pollution")
     op.drop_column("station", "wording", schema="transport")
@@ -31,7 +46,6 @@ def upgrade() -> None:
 
 def downgrade() -> None:
     """Downgrade schema."""
-    op.drop_column("time_bin", "cat_day", schema="transport")
     op.add_column(
         "station", sa.Column("wording", sa.String(), nullable=True), schema="weather"
     )
@@ -41,3 +55,12 @@ def downgrade() -> None:
     op.add_column(
         "station", sa.Column("wording", sa.String(), nullable=True), schema="transport"
     )
+
+    op.execute("""
+        ALTER TABLE transport.time_bin 
+        ALTER COLUMN cat_day TYPE VARCHAR 
+        USING cat_day::text
+    """)
+
+    op.drop_column("time_bin", "cat_day", schema="transport")
+    op.execute("DROP TYPE IF EXISTS transport.cat_day")