+++ public_transport_watcher/db/models/enums/__init__.py
@@ -1,0 +1,5 @@
+from .day_category import DayCategoryEnum
+
+__all__ = [
+    "DayCategoryEnum",
+]