+++ public_transport_watcher/extractor/configuration/extraction_config.py
@@ -1,5 +1,5 @@
 EXTRACTION_CONFIG = {
-    "batch_size": 10000,
+    "batch_size": 1000,
     "navigo": {
         2022: ["s1", "s2"],
         2023: ["s1", "s2"],